requirements.txt
environs==14.1.1
python-telegram-bot==13.15
redis==5.2.1